+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 **?.?.? Changes:**
 
+* Added config options to prevent the mod from disabling attack collisions with/from specific characters
+
 * Fixed player projectiles still impacting allies if a player changes to another team.
 
 **1.0.1 Changes:**